Block users frontend

+++ backend/src/main/java/Nuvola/Projekatsiit2025/repositories/RegisteredUserRepository.java
@@ -4,6 +4,6 @@ import Nuvola.Projekatsiit2025.model.RegisteredUser;
 import Nuvola.Projekatsiit2025.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RegisteredUserRepository extends JpaRepository<RegisteredUser, Integer> {
+public interface RegisteredUserRepository extends JpaRepository<RegisteredUser, Long> {
     RegisteredUser findByUsername(String username);
 }
